update: 优化代码
