+++ src/hooks/index.ts
@@ -1,0 +1,5 @@
+import useQuestion from './useQuestion'
+
+export {
+  useQuestion
+}
